requirements.txt
tornado>=3.2
torndb>=0.2
PyYAML>=3.10
pycrypto>=2.6.1
beautifulsoup4>=4.3.2
lxml>=3.3.2
MySQL-python>=1.2.5
